+++ src/main/java/org/bytedeco/javacpp/tools/ParseMojo.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (C) 2018 Samuel Audet
+ *
+ * Licensed either under the Apache License, Version 2.0, or (at your option)
+ * under the terms of the GNU General Public License as published by
+ * the Free Software Foundation (subject to the "Classpath" exception),
+ * either version 2, or any later version (collectively, the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.gnu.org/licenses/
+ *     http://www.gnu.org/software/classpath/license.html
+ *
+ * or as provided in the LICENSE.txt file that accompanied this code.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.bytedeco.javacpp.tools;
+
+import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugins.annotations.LifecyclePhase;
+import org.apache.maven.plugins.annotations.Mojo;
+
+/**
+ * A {@link BuildMojo} that always forces {@link BuildMojo#generate} to false,
+ * with the goal of parsing header files instead.
+ *
+ * @author Samuel Audet
+ */
+@Mojo(name = "parse", defaultPhase = LifecyclePhase.GENERATE_SOURCES)
+public class ParseMojo extends BuildMojo {
+    @Override public void execute() throws MojoExecutionException {
+        generate = false;
+        super.execute();
+    }
+}